+++ src/daf/__init__.py
@@ -22,7 +22,7 @@ import sys
 import warnings
 
 
-VERSION = "3.3.0"
+VERSION = "4.0.0"
 
 if sys.version_info.minor == 12 and sys.version_info.major == 3:
     warnings.warn(