+++ src/daf/messagedata/__init__.py
@@ -1,2 +1,3 @@
 from .basedata import *
 from .textdata import *
+from .voicedata import *