+++ src/daf/messagedata/voicedata.py
@@ -1,13 +1,14 @@
 from typing import List, Callable, Any, Dict, Coroutine
 from dataclasses import dataclass, asdict
+from contextlib import suppress
 
 from ..dtypes import FILE
 from .basedata import BaseMessageData
 from ..misc.doc import doc_category
-from ..misc.async_util import except_return
+from ..logging.tracing import trace, TraceLEVELS
 
 
-__all__ = ("VoiceMessageData", "DynamicVoiceMessageData")
+__all__ = ("BaseVoiceData", "VoiceMessageData", "DynamicVoiceMessageData")
 
 
 @doc_category("Message data", path="messagedata")
@@ -45,13 +46,25 @@ class DynamicVoiceMessageData(BaseVoiceData):
         self.args = args
         self.kwargs = kwargs
 
-    @except_return
     async def to_dict(self) -> dict:
-        result = self.getter(*self.args, **self.kwargs)
-        if isinstance(result, Coroutine):
-            result = await result
+        try:
+            result = self.getter(*self.args, **self.kwargs)
+            if isinstance(result, Coroutine):
+                result = await result
 
-        if isinstance(result, VoiceMessageData):
-            return await result.to_dict()
+            if isinstance(result, VoiceMessageData):
+                return await result.to_dict()
+            elif result is not None:
+                # Compatibility with older type of 'data' parameter. TODO: Remove in future version (v4.2.0).
+                if not isinstance(result, (list, tuple, set)):
+                    result = [result]
 
-        return None
+                file = None
+                if isinstance(result[0], FILE):
+                    file = result[0]
+
+                return await VoiceMessageData(file).to_dict()
+        except Exception as exc:
+            trace("Error dynamically obtaining data", TraceLEVELS.ERROR, exc)
+
+        return {}