+++ src/daf/misc/instance_track.py
@@ -48,13 +48,18 @@ class ObjectReference:
         return cls(get_object_id(obj), obj._tracked_allow_remote)
 
 
+TRACK_DEFAULT_ATTR = {
+    "_daf_id": -1,
+    "_tracked_allow_remote": True
+}
+
 def track_id(cls):
     """
     Decorator which replaces the __new__ method with a function that keeps a weak reference.
     """
     @wraps(cls, updated=[])
     class TrackedClass(cls):
-        if hasattr(cls, "__slots__"):  # Don't break classes without slots
+        if hasattr(cls, "__slots__") and cls.__slots__:  # Don't break classes without slots
             __slots__ = ["_daf_id", "_tracked_allow_remote"]
             if not hasattr(cls, "__weakref__"):
                 __slots__.append('__weakref__')
@@ -86,8 +91,10 @@ def track_id(cls):
 
         def __getattr__(self, __key: str):
             "Method that gets called whenever an attribute is not found."
-            if __key == "_daf_id":
-                return -1
+
+            value = TRACK_DEFAULT_ATTR.get(__key)
+            if value is not None:
+                return value
 
             raise AttributeError(__key)