+++ src/daf/responder/__init__.py
@@ -7,3 +7,4 @@ from .guildresponder import GuildResponder
 
 from .constraints import *
 from .actions import *
+from .logic import *