+++ testing/fixtures/__init__.py
@@ -1,0 +1,4 @@
+from daf_gui.tod_extensions.loader import register_extensions, register_annotations, register_conversions
+register_extensions()
+register_annotations()
+register_conversions()